requirements.txt
Django==1.5.1
Markdown==2.3.1
distribute==0.6.27
django-extensions==1.1.1
django-filter==0.6
djangorestframework==2.3.6
ipython==0.13.2
readline==6.2.4.1
six==1.3.0
wsgiref==0.1.2
